project/requirements.txt
sounddevice
soundfile
openai
whisper
streamlit
python-docx

